+++ libft/mandatory/part1/ft_atoi.c
@@ -6,17 +6,19 @@
 /*   By: rolee <rolee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/09 22:16:24 by rolee             #+#    #+#             */
-/*   Updated: 2024/03/20 13:26:56 by rolee            ###   ########.fr       */
+/*   Updated: 2024/04/24 20:33:33 by rolee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../libft.h"
 
-int	ft_atoi(const char *str)
+#include <stdio.h>
+
+long long	ft_atoi(const char *str)
 {
-	int	idx;
-	int	sign;
-	int	num;
+	int			idx;
+	int			sign;
+	long long	num;
 
 	idx = 0;
 	sign = 1;
